chore: truncate response text when error is not JSON
This is probably unnecessary, the cases where the logging was too large
was because of printing the raw response, which includes the request too
and for huge files it's printing those huge files that caused problems.
Now that we no longer print/return the raw response we shouldn't
encounter those issues anymore, but truncating the response is ok as a
precaution in case it's extremely large.